+++ extension/webview/src/components/SlashCommandDropdown/index.ts
@@ -1,0 +1,1 @@
+export { SlashCommandDropdown } from './SlashCommandDropdown';